+++ src/shared/theme-preset-assets.manifest.js
@@ -1,0 +1,9 @@
+export const THEME_PRESET_ASSET_FILES = Object.freeze({
+  "british-flag": "british-flag.jpg",
+  cyberpunk: "cyberpunk.jpg",
+  matrix: "matrix.png",
+  fire: "fire.jpg",
+  ice: "ice.jpg",
+});
+
+export const THEME_PRESET_ASSET_KEYS = Object.freeze(Object.keys(THEME_PRESET_ASSET_FILES));